requirements.txt
python-bitcoinlib>=0.11.0
requests>=2.25.1
python-dotenv>=0.19.0 